+++ php/src/Indexing/BuiltinIndexer.php
@@ -344,8 +344,8 @@ class BuiltinIndexer
         $structureTypeMap = $this->getStructureTypeMap();
 
         $structureId = $this->storage->insertStructure([
-            'name'              => $element->getShortName(),
-            'fqcn'              => $element->getName(),
+            'name'              => $this->getStructureShortName($element),
+            'fqcn'              => $this->getStructureFqcn($element),
             'file_id'           => null,
             'start_line'        => null,
             'end_line'          => null,
@@ -483,6 +483,53 @@ class BuiltinIndexer
         $this->storage->update(IndexStorageItemEnum::CONSTANTS, $constantId, [
             'structure_id' => $structureId
         ]);
+    }
+
+    /**
+     * Retrieves the short name of the specified classlike.
+     *
+     * Some of PHP's built-in classes have inconsistent naming. An example is the COM class, which is called 'COM'
+     * according to the documentation but is actually called 'com' when fetching a list of built-in classes. PHP in
+     * itself is mostly case insenstive, but as we aren't, we must at least ensure that the users get the expected
+     * results when consulting the documentation.
+     *
+     * @param ReflectionClass $element
+     *
+     * @return string
+     */
+    protected function getStructureShortName(ReflectionClass $element)
+    {
+        $correctionMap = [
+            'com'     => 'COM',
+            'dotnet'  => 'DOTNET',
+            'variant' => 'VARIANT'
+        ];
+
+        $shortName = $element->getShortName();
+
+        return isset($correctionMap[$shortName]) ? $correctionMap[$shortName] : $shortName;
+    }
+
+    /**
+     * Retrieves the FQCN of the specified classlike.
+     *
+     * See {@see getStructureShortName} for more information on why this is necessary.
+     *
+     * @param ReflectionClass $element
+     *
+     * @return string
+     */
+    protected function getStructureFqcn(ReflectionClass $element)
+    {
+        $correctionMap = [
+            'com'     => 'COM',
+            'dotnet'  => 'DOTNET',
+            'variant' => 'VARIANT'
+        ];
+
+        $shortName = $element->getShortName();
+
+        return isset($correctionMap[$shortName]) ? $correctionMap[$shortName] : $shortName;
     }
 
     /**